pyproject.toml
[project]
name = "kissllm"
version = "0.0.1"
description = "Python SDK to call all LLM apis."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langfuse>=2.59.3",
    "openai>=1.63.0",
    "mcp>=0.1.0",
    "asyncio",
]

[dependency-groups]
dev = [
    "python-dotenv",
    "pytest>=8.3.4",
    "pytest-asyncio",
]
